Vanilla bugfix, C343-149: `spestudied' manipulation for polymorphed spellbooks
From a bug report... some special case handling for polymorph of
spellbooks never worked as intended.  It's possible to polymorph a
spellbook, use it to learn a new spell, and then repeat as many times
as you like unless/until you run out of polymorph magic or the small
chance of "object shuddering" causes it to be destroyed.  Polymorph
was incrementing the book's ``number of times read'' field with the
intent that it would fade to blank after being read 3 times (which
turns out to the 4 times since the check is actually for re-reading 3
times after the first).  That didn't work because the spestudied field
was ignored when learning a new spell, only checked when relearning a
known spell.

Now it will be checked when learning a new spell, and also the book
tweaking during polymorph is slightly more elaborate.  If you happen
to get a blank book during the item selection, it will have a read
counter of 0 and can be re-polymorphed into something readable; but if
you get some other book, its read counter will be set to one greater
than than the original book's (same as before), and then the counter
will be checked to see if it has gone over the limit, in which case
the book will be made blank and its counter will be reset to a random
value.

Re-polymorphing that blank book again has 1/4 chance apiece among the
following cases:
 * book gets blanked again; goto step 1...
 * book is non-blank but too faint to read; reading attempt will fail
 * book can be read normally and then re-read once
 * book can be read normally and then re-read twice
which is more inline with the intent of the original special case
code.  It's actually slightly nastier since you'll occasionally get a
book for a spell you don't know yet but then not be able to learn it
from that book.

(cherry picked from commit 98685ea23479afe732a9047fe5e55b059adc0795)

Conflicts:
	doc/fixes35.0
	include/hack.h
	src/spell.c
	src/zap.c