Balance adjustment for death drops.
The probability of a death drop is now inversely proportional to the
kill count for the type of monster in question.  The first few
monsters of any given type that you kill are now more likely to drop
items than previously, but the probability quickly drops off as you
kill more and more of that type of monster.  The drop-off is on the
same kind of logarithmic scale as the corpse dropping but is
calculated independently (because the decisions are made in different
places in the code).