Elbereth balance:  when writing/engraving on the floor, the chance for each character to be written differently than intended are directly proportional to the number of times you have written Elbereth (as counted for conduct purposes) and inversely proportional to your experience level.  The odds are best (lowest chance of mistakes) when burning, worst when writing in dust or blood.  Players who use Elbereth in emergencies will find it (especially dust Elbereth) noticeably more reliable than in 3.4.3 (the first time you use it, you have a very high probability of getting it written successfully on the first try), but Elbereth is now less susceptible to extreme abuse. "Spamming" it extensively will just give you hand cramps.
As a result of this change, it will probably be necessary to rebalance
the rate at which various actions erode engravings, but that can be
done in a separate commit.

Another side effect will be that athames are likely somewhat less
useful.  This will likely lead to the Wizard role needing some small
bit of compensation (for Magicbane being less powerful, especially in
the early game), but again, that can be handled separately.  Actually,
upstream changes may take care of that in 4.4 timeframe, if the
proposed wisdom-based power regeneration is enacted as discussed under
Class Overhaul.