Filename changes.
I got tired of having 4 and Fourk putting their save and config files
in the same directory, so I redid the aimake rules changes and appear
to have gotten them more or less correct this time (at least, enough
of them to solve the immediate problems I was having).  While I was
at it, I fixed some references to the NH4 filenames in some of the
documentation.

This commit also contains a minor tweak to the menu screen logo.