Ticket #613, Fix C343-295, Casting stone-to-flesh on yourself while wielding a figurine can cause the game to panic.
The symptom was slightly different in NH4 from 3.4.3, but the cause
was the same:  the figurine was still wielded when it was deleted (and
thus freed).  Rather than fix the bug in the stone-to-flesh code in
zap.c, I instead put the fix in delobj() and made it somewhat more
general.  Now any time delobj() is called on an object that is wielded
or worn, it will be unwielded or unworn _before_ it is actually deleted.

This is different from the Dev Team's fix, which does not apply
cleanly and easily to NH4 due to other code differences.  (However,
the Dev Team's fix is still worth looking at, because it also fixes
other stone-to-flesh bugs, related to golem statues and figurines,
which I have not attempted to address in NH4 yet.)