Make xorns red and acid blobs yellow.
Eventually, xorns will be on the same letter as umber hulks, so it
would be good if they weren't both brown.  Similarly, blobs will
probably end up on the same letter as jellies, so it would be good if
the color green weren't used twice -- and yellow is in any case
consistent with yellow dragons.  I'm not going to make all acidic
monsters yellow (among other things, there are too many of them, and
some have the colors that they have for good reasons), but acid blobs
actually have acid in their name, and acid is the major thematic
feature of yellow dragons.  So those two are now the same color.