Apply fix for C343-9 from nethack.rankin Sat Dec 27 02:45:47 2003.
This is the Dev Team's fix for the bug wherein the incorrect article
is used for a trapped steed when the player is hallucinating, as in,
"You lead poor [hallumon] into a pit."

The Dev Team's patch also did some code refactoring for whitespace and
to rename the mtmp variable to steed in the steedintrap function.  The
former is moot for NH4 due to our holistic re-indentation; the latter
change would probably be a good idea, but I contented myself with just
importing the actual bug fix for now.