Apply mbodypart fix from Dev Team, michael.allison, Sun Feb 19 21:57:54 2006 +0000
This patch adds worm_parts, as distinct from snake_parts (C343-157);
gives krakens tentacles (C343-158); and ensures that stalkers have a
head (C343-159).

The stalker fix may be redundant in NH4 (Ticket #434 claims that bug
had already been fixed), but if so it was fixed in another section of
the code.  I went ahead and applied the Dev Team fix anyway, including
the stalker portion; it applies cleanly and does not appear to cause
problems.  At worst it's a single unnecessary redundant conditional.