Make candles in the player's inventory subject to fire damage in situations where other objects would be.
In situations where scrolls, books, and potions that you are carrying
can be destroyed by fire, candles can now be set afire as well, which
uses up some of their remaining wax (or tallow) and leaves them lit.

Apart from filing a ticket (#800), I did NOT attempt to address the
numerous situations where it would make logical sense for flammable
items to suffer damage from fire but currently that is not the case.
The goal at this point was just to bring candles to a reasonable level
of parity with other flammable items, for ticket #432/C343-154.

I also didn't deal with the case where the candles are already lit.
Logically, fire damage to an already lit candle should reduce its
amount of fuel remaining, but that would require adjusting existing
timers, which (because of the way run_timers works) would require
sorting the list of timers afterward.  I did not attempt that.  My
knowledge of C in general and C pointers and memory management in
particular being as limited as it is, there is a non-trivial
probability I would screw it up in a way that could result in
significantly worse bugs.  So for the time being lit candles are
immune to non-lava fire damage (other than the tiny amount per turn
that they suffer simply by virtue of being lit).