Ticket #435, C343-162, apply Bilious patch #445, from Ray Chason.
This patch provides singular verbs ("is" or "resists") when a monster
is stunned or confused by being hit with Magicbane.  This patch uses
youdefend to determine the correct verb tense (always selecting
third-person singular for monsters other than the player), rather than
calling vtense at all.

I've looked at the Dev Team's fix for this (which does still use
vtense, except when the entity being hit is named), but there are two
problems with it:  it only fixes "is" to "are" (failing to fix
"resist" to "resists"), and it doesn't apply straightforwardly to NH4
because of significant code divergence (having to do with mextra
fields -- NH4 doesn't appear to even have some of the data structures
involved, much less the mextra.h macros that the 3.4.3 Dev Team's fix
relies upon).  So I went with the Bilious patch for this one.