Ticket #603, C343-287, apply Dev Team fix, nethack.rankin Mon Apr 16 03:03:08 2007 +0000
If polymorphed into a monster with a kick attack, and you either
explictly kicked a monster with Ctrl-D or attacked it in the normal
melee fashion getting the "You kick ..." message, you did not receive
the damage bonus from rings of increase damage.  Although obscure in
normal play (because damage inflicted upon others does not have its
numbers directly disclosed to the player), this could be trivially
verified in wizard mode using a stethoscope and a highly enchanted
ring of increase damage.

Once noticed, it seemed clearly wrong, because you DID receive the
bonus when kicking if NOT polymorphed into a monster with a kick
attack, so it seemed that non-kicking monster forms were actually
better at kicking than forms that were supposed to be good at it.