Ticket #34, C343-75: Apply Dev Team fix, nethack.rankin Thu Feb 3 04:12:28 2005 +0000
This fixes the issue wherein food detection does not warn that
Medusa's corpse is dangerous to eat if you have poison resistance,
because touch_petrifies() returns false.  The fix introduces a new
function in eat.c, flesh_petrifies, which returns true if eating that
monster will petrify you.

The Dev Team's fix also handles the hypothetical case of a Medusa egg,
but I didn't apply that portion because the relevant code has diverged
too much (between changes to the dev team's code and changes to NH4)
for it to apply cleanly.  It isn't needed in vanilla anyway, since
Medusa isn't an egg-laying monster.  If a variant introduces a new
monster that lays eggs and causes petrification when eaten but not
when touched, this issue will then need to be revisited.