Apply Dev Team fix for swallowing Medusa and such, nethack.rankin Sat Dec 3 03:16:25 2005 +0000
If eating Medusa's corpse turns you to stone, swallowing her whole
with a digestion attack should do so as well.

Extra bonus fix: dying by swallowing whole (with a digestion attack) a
monster that is fatal to eat now violates foodless conduct (and also
vegetarian and vegan if eating the monster in question isn't in line
with those conducts).