Ticket #35, C343-76, Apply food appraisal when eating tins.
The edibility_prompts() function has code for handling tins, but
doeat() wasn't calling edibility_prompts() when the player eats a tin.
Now it will.  Besides Medusa, this is also relevant for tins of
cockatrice, chickatrice, and I think green slime.  Not sure if there
are other dangerous tins beyond that, but in any case I didn't change
the edibility_prompts() logic at all, just adjusted the tin-eating
code so that it actually calls it the function and aborts if the
player chooses not to eat.