Apply Dev Team fix for C343-201, nethack.rankin, Tue Aug 8 05:08:23 2006 +0000
A beta tester had reported that a monster wielded a stack of spears
and then thrust "its spear" at another monster.  This patch introduces
to mhitm the "one of" wording.  The commit message implies that this
wording was already in place for the mhitu case; I will investigate
separately whether that is also the case in NH4.