Apply Dev Team fix, SC343-14, nethack.rankin, Thu Aug 10 00:45:32 2006 +0000
A beta tester had reported that under some obscure conditions, a
temple priest might abandon his temple via teleport.  Similar logic
was found to be relevant also to shopkeepers.  The fix changes
rloc_pos_ok() so that for shopkeepers and priests it will only accept
spots inside the priest's temple or inside the shopkeeper's shop IF
they are currently in that room.

For direct attacks from the player, this only matters for shopkeepers
(priests will resist), but a priest can be teleported via indirect
means (e.g., quantum mechanic and conflict) and will now try to select
a position still within the temple.  Also, the original commit comment
suggests that a priest polymorphed into something like a nymph would
use the new behavior when warping.

If the calling code runs through the max number of retries, it still
reverts to goodpos() as before, so in pathological cases it will
approximate the 3.4.3 behavior.

Also as before, shopkeepers or priests who get polymorphed into
something covetous will still warp to the upstairs when injured.