Projectile weapons buff: arrows and darts.
All types of arrow now do one additional point of base damage against
small monsters (no change against large).

Darts do one additional point of base damage across the board.

This buff is intended to make these pure projectiles a slightly more
attractive option, especially in the early game, and particularly as
compared to daggers.  By itself, this change is probably inadequate,
but together with other proposed changes (e.g., when the roles are
overhauled, fewer roles will be able to multi-fire daggers) it should
help to rebalance things so that daggers are not obviously the final
word in ranged weapon strategy for all situations.

It is understood that this buff will likely benefit certain roles
(most notably Tourist, but also Ranger and perhaps to some extent
Samurai).  Hopefully it won't be too unbalancing, but we'll see.