Apply Dev Team fix for C343-220, nethack.rankin, Sat Oct 7 05:41:30 2006 +0000
A player had reported that when attempting to eat a rider corpse
(which is ordinarily fatal), there was a 1/7 chance that the corpse
would be randomly treated as rotten.  If this happened and the "world
spins" message was printed, the player survived and ended up with a
partly eaten rider corpse.

The fix treats rider corpses like lizards and lichens for rotten-food
purposes, ensuring that they are always fatal to eat.  The
partly-eaten corpse result can still be achieved via life-saving,
but the corpse will immediately revive in that case.

Added bonus fix: lembas wafers and cram rations can only be rotten if
they are cursed.