Apply Dev Team fix for C343-221, nethack.rankin, Fri Oct 13 04:03:36 2006 +0000
It had been mentioned in rgrn that monsters that can be generated
asleep and can also be steeds (the given example was a Jabberwock) can
be saddled, mounted, and ridden while asleep.  Movement was checking
for timed sleep but neglected to check for the variety where the
monster is waiting for something (or someone) to wake it up.

The fix moves the check into a function, which checks for both kinds
of sleep and can optionally also check for eating.

Additionally, monsters now have a chance to wake up when someone
attempts to saddle or ride them.  Even if they don't wake up outright,
their remaining sleep turn count (if applicable) may be reduced.