Apply Dev Team fix for C343-222, nethack.rankin, Fri Oct 13 05:39:16 2006 +0000
It was reported on rgrn that wearing life-saving and wielding a
cockatrice corpse bare-handed while polymorphed into a form that can
do so, if you rehumanized and were life-saved from the stoning, could
result in still wielding the corpse without being stoned again.

In principle, the same thing could happen to a monster.

The fix causes the corpse to become unwielded if the wielder gets
life-saved, for both players and monsters.

Not included here, there's also a related fix for a stone-to-flesh
case.  I will look at that separately and, if applicable to NH4, add
it as a separate commit.