Apply Dev Team fix, stone-to-flesh case related to C343-222, nethack.rankin, Sat Dec 2 04:44:56 2006 +0000
If the player is wielding a cockatrice corpse without gloves while
polyselfed into a stone golem, stone-to-flesh could result in the
player still wielding the corpse in flesh golem form.

The fix makes the corpse turn the player back into a stone golem,
once the first transformation is complete.

Added bonus fix:  polyself was meant to exercise constitution, but
because exercise() was called in the wrong place, it had no effect
(because exercising most attributes has no impact while polyselfed).
This commit moves the exercise() call to the correct place in the
function, so that it has the intended effect.