Remove lemures.  They added absolutely nothing of consequence to the game, just needlessly cluttered up the monster list for no reason.
I've left manes in because they can occur early enough in the game
that they are, at that point, one of very few normal-speed monsters to
be a non-threat, which is potentially interesting in itself.  By the
time lemures can occur, on the other hand, there are large numbers of
normal-speed and even fast types of monsters that are completely
harmless to the player, so adding yet another one that late in the
game is just pointless.

The three hardcoded lemures in gehennom.des are converted to two manes
and an imp; not that any of the i are really meaningful post-castle,
but the whole of Gehennom is going to get majorly overhauled anyway
at some point.