Apply Dev Team fix for C343-224, nethack.rankin, Thu Nov 2 06:26:37 2006 +0000
If the player is carrying three or more items in inventory slot #,
the inventory letter list for any "What do you want to [verb]?"
prompt could contain a # for each, which was pointless because
you can't select them individually anyway.  This compacts any
string of three or more # symbols in such a list to #-# range,
in much the same way that abcdef becomes a-f.

This patch does not address the cause, i.e., it makes no attempt
to prevent the player from getting multiple items into the # slot.
It just improves compatification of the resulting prompts.

In 3.4.3, the easiest way to reproduce the issue is to polyself into a
giant and pick up a bunch of boulders.  I'm not sure if that
particular approach is still possible in NH4 (there's another bug on
the list for that -- player can pick up too many boulders, or
something along those lines -- and I don't recall whether we've fixed
it yet), but if not you can always fill your regular inventory
letters, then still pick up one boulder, and then I believe you can
still pick up some gold into a second # slot, then pick up loadstones
of different beatitudes, for a total of up to five # slots.  This
patch will turn the resulting ##### into #-# in prompts.