Apply Dev Team's partial fix for C343-225, nethack.rankin, Fri Nov 3 04:41:30 2006 +0000
This limits the number of boulders and/or loadstones you can pick up
when you're out of inventory slots.  You can always pick up one (if
polymorphed into a giant, in the case of a boulder), but once you're
already carrying some the game will now stop you if you don't have
inventory slots avaialble.

It is believed to still be possible to get up to three items into
overflow-inventory slots (#): one other item, plus a boulder, plus
a loadstone.