Apply (most of) first Dev Team fix for UNL-031, nethack.rankin, Tue Nov 7 02:49:09 2006 +0000
I didn't attempt to apply the portion related to #tip, since NH4 does
not have that feature at this time.

Also, one thing will very likely need to be adjusted.  The pick_pick
function (which I've named pick_pick_from_container to be slightly
less obscure) contains a static variable, which is probably
incompatible with the NH4 save system and likely could cause desyncs.
We'll probably need to move that variable to program_state or
someplace.  I put in a comment to that effect but haven't had time to
decide what to actually do yet, so this is not ready to merge.

Additionally, the Dev Team had a second related commit, on
Tue Nov 7 05:19:48 2006 +0000, which I should look at before merging.