Apply second Dev Team commit for UNL-031, nethack.rankin, Tue Nov 7 05:19:48 2006 +0000
This patch is mainly focused on the use of the "cad" insult, which
apparently is indicated in some dictionaries for use only against male
targets.  The fix introduces a cad() function, which returns different
insults depending on the demon-ness and gender of the player character.

I've taken the liberty of removing the unnecessarily complex alternate
usage case, which was only used once, from the cad() function, leaving
the punctuation in the caller's string where it clearly belongs and
using msgupcasefirst(cad()) to effect capitalization.

This still isn't quite right for NH4, partly for reasons noted in the
previous commit message (q.v.), so it still needs to be adapted
somewhat before merging.