Fix up the UNL-031 fixes to fit NH4 (globals purge and deduplication)
The major thing here is moving a static variable to program_state.

The other thing is, I'd incorrectly thought we didn't have the
inhishop function yet and copied it in to support one of the fixes
that uses it; we already did, in fact, already have it, so this
removes the duplicate function definition.

With these fixes, the code now compiles and runs, and I've tested
several cad-message cases successfully.