Apply Dev Team fix for UNL343-033, nethack.rankin, Thu Nov 23 03:06:19 2006 +0000
A player had reported that entering a magic trap on the same turn
levitation times out causes the trap to trigger twice in a row.  In
fact, it only seemed that way if no monsters attacked ad interim, but
it really did seem weird in that case.

The fix adjusts the levitation timeout by one turn, either forward or
backward with equal probability.  So either you float down on entering
the tile and the trap triggers immediately (but just once), or else it
gets delayed for an extra turn so you get a chance to act before you
float down and trigger the trap again.

This fix applied to NH4 almost without modification.