Apply first Dev Team fix for UNL343-034, nethack.rankin, Thu Nov 23 03:32:53 2006 +0000
High monster MR made the other three riders unlikely to be polymorphed
as such, but they were vulnerable to being slimed (and would not revive
if killed in that state).  This fix makes them completely immune to
both kinds of transformation.

This is slightly asymetrical, since War is not inherently immune to
such things, but by the time most players reach the Astral plane they
may as well be, so ok; for gameplay reasons, it is clearly not
desirable that the other three riders be easily eliminated this way.