Apply second Dev Team fix for UNL-034, nethack.rankin, Thu Nov 23 04:17:48 2006 +0000
This one makes it slightly harder to eliminate the Riders by filling
the astral plane with monsters.  In 3.4.3, if the level is full when
the game tries to revive any corpse, it rots away instead.  Now, if
the corpse is a rider corpse, it has a 99% chance to be tried again
a few turns later.

In practice, I'm not sure if this will actually make any difference,
since once the level is full the player can just leave it full until
the less hassled message is eventually received.  But from a flavor
perspective it's nice that the Riders don't rot away so quickly.

The fix also applies the M3_DISPLACES flag to monster revival, and so
I have gone ahead and imported the flag to NH4.  However, I didn't
implement it for monster movement, just the revival code contained in
the fix.  If we later decide that we want live riders to shove other
monsters out of their way when moving, we can import those changes
at that time -- perhaps for version 4.4 we might consider this.