Correct two uses of mon->mnum to monsndx(mon->data).
Turns out mnum was actually the original mnum from when the monster
was generated, not its current form.  It was not updated when the
monster grew up (e.g., when your kitten became a housecat) nor yet
when it became transformed due e.g. to polymorph, and so it was
unreliable for most purposes.

This general category of error was discovered a couple of days ago
upstream and corrected there by Alex Smith, but previously I had
already introduced two additional instances in Fourk.  This commit
corrects those two additional instances, which were Fourk-specific.

We can be pretty sure we've got all the instances, because Alex went
ahead and renamed mnum to something less likely to be misunderstood,
and so any remaining references to it would block compile.