Apply Dev Team fix for C343-227, nethack.rankin, Sun Nov 5 04:16:10 2006 +0000
When an item is dropped in a shop and not sold, its no_charge bit is
set.  When the item is subsequently picked up at no charge, this is
cleared (otherwise, dropping the same item in a different shop and
selling it would leave the item player owned, and it could be picked
up and dropped again and sold a second time).  Provided the first shop
is still tended when the item is picked up, this was happening
recursively for container contents, as it should; but if the first
shop had become untended after the no_charge bits were set, then
container contents would be left with their no_charge bits still set,
which would allow them to be sold in a second shop and still remain
player owned (once; or sold and picked up and sold a second time).

This fix causes container contents to have their no_charge bit
cleared, recursively, when the player picks them back up, even if
the shop is now untended.