Apply Dev Team Fix, C343-230, nethack.rankin, Thu Nov 30 02:18:49 2006 +0000
If the player zapped certain wands or spells up or down while standing
on the portcullis directly to the north of an open drawbridge, the
game was interpreting the return value of is_drawbridge_wall(), which
was 0 for north, as being 0 for false as would be expected with a
purely boolean function.  In fact, is_drawbridge_wall returns a
negative value for false.

The fix corrects the conditional to check whether the return value is
>= 0 rather than whether it is true.