Rename is_drawbridge_wall() to drawbridge_wall_direction().
The function returns a direction (DB_NORTH, DB_SOUTH, DB_EAST, or
DB_WEST), so the name of the function should reflect that.

Yes, it returns a fifth possible value (-1) in the event that the
location is not a drawbridge wall at all, and yes, it is frequently
used just to check for this, but the misleading name has already
caused at least one bug, and in general any programmer could be
excused for expecting a function named is_drawbridge_wall() to return
the answer to whether the location is a drawbridge wall (true, one
would expect) or not (false, presumably), which is not the case.  This
commit heads that off at the pass, hopefully.

The name of is_db_wall() could probably stand to be clarified as well,
and perhaps we should audit all the place it's called to make sure
they weren't meant to be the other function, but I haven't gone that
far in this commit.