Apply first Dev Team fix for C343-236, nethack.rankin, Thu Dec 7 06:28:16 2006 +0000
It was reported on the newsgroup that if polymorphed into a phasing
form you could engrave in places that don't really have a surface to
write on, such as solid rock.  There was a patch offered in the
newsgroup, but it didn't handle all of the possibilities correctly
(e.g., trees, iron bars), so the Dev Team's fix is more general.  I've
applied it almost exactly verbatim, sans a pline-wrapper function
optimization that Nitro-derived variants don't use.

There's another related commit from a few days later that I still need
to look at (wherein the accessible() function that this code uses had
a bug that is relevant both here and elsewhere), and additionally
while testing this I discovered that you can write in the floor while
embedded in it, so I want to have a look at that too.