Apply second Dev Team fix related to C343-236, nethack.rankin, Fri Dec 15 02:36:58 2006 +0000
This fixes a bug in a number of related areas of code that were
checking to see if a certain tile on the map was accessible.  The case
that got it noticed had to do with engraving in inaccessible
locations: the fix for C343-236 made this check use the accessible()
function, which introduced an unintentional change related to raised
drawbridges.  It turned out that a number of other places in the code
were also failing to handle that situation correctly, by failing to
check the underlying terrain.

This puts the check for the underlying terrain into accessible()
and makes a number of checks call that function.  I should probably
audit all places ACCESSIBLE are called in case there are any more.