Adjust comment on ACCESSIBLE to refer to accessible().
The ACCESSIBLE macro is still needed, partly because accessible()
calls it but mostly because it is also used in the vision code, which
I am not confident changing.  However, I did look at all the places
where ACCESSIBLE is used directly, and I don't think any of the
remaining ones should be subject to the bug that the recent changes to
accessible() were designed to fix.  For example, I would hope that
vault guards would not make their special corridors across drawbridges
even if they could occur on the same level as one, which currently is
not the case.  It might be possible to do some additional future
proofing here, but for now this comment change should be adequate.