"You can't write in the [floor] while embedded therein."
If you sink into lava then freeze it by zapping cold down, this leaves
you embedded in the floor.  Previously you could then write on the
floor, just as if you were standing on it like usual.  Now you can't.

I wanted to make the message say "Your [body] is in the way", but
body_part doesn't seem to have a constant for the entire body, unless
I missed it, so I went with the message above for now.