Apply Dev Team Fix, UNL-039, from nethack.rankin, Sun Dec 17 06:26:08 2006 +0000
If a monster died while multi-firing (e.g., because one of the fired
items hit a gas spore that exploded), and if when the code that killed
the monster dropped its inventory on the floor this resulted in the
items being thrown merging into an existing on-the-floor stack, the
memory referenced by the items-being-thrown variable would be freed.
Continuing to reference this memory as the "remaining" in-the-air
items were being processed by the throwing code could then result in
undefined behavior.

In principle, the ideal fix would involve all the individual
multi-fired items being treated as in the air simultaneously so that
when the monster dies, its other inventory would drop to the floor but
the items being fired would not and thus could not merge until after
the throwing code finishes with them.

This fix doesn't go that far.  For simplicity, it simply cancels the
remainder of the volley if the monster dies.  Since monsters don't
very often die while firing volleys (while something else is firing a
volley yes, but not so often when the monster itself is doing so),
this is unlikely to have a major impact on game play or balance.