When the drawbridge was raised or lowered, the engraving remained, so from the player's in-game perspective it appeared to transfer between the bridge and the terrain underneath the bridge.
The ideal fix would be to have distinct engravings (or lack thereof)
for the bridge and the terrain beneath it, but this fix does not
implement that.  It just causes the act of raising, lowering, or
destroying the bridge to remove any engravings, as it already does
for traps.

While I was testing this fix for NH4, I discovered that the message
for engraving with a wand of digging seems kind of inappropriate if
you do it on a drawbridge.  That may also be the case for some other
locations.  I'll look into that as a separate commit.