Adjust the message when engraving on a drawbridge with a wand of digging.
Discovered when testing the fix for C343-239.  "Gravel flies up from
the floor" didn't feel quite right.  I asked on #nethack-dev, and
bcode suggested "Splinters fly up from the floor."  I settled on
"Splinters fly up from the bridge."

There's potentially a latent minor message bug here, which I did not
address, that could manifest if drawbridges are introduced on
bones-eligible levels, levels with random graves, or levels with
graves under the bridge for some reason.  See the changes to the
accessible() function that were rolled in a handful of commits ago for
inspiration on how to deal with it.  Note that the ice case (which
*does* occur if the player freezes the moat) is already handled
appropriately because is_ice already checks for a raised drawbridge
with ice under it, so the case that is currently actually reproducible
is covered.  The grave case uses the IS_GRAVE macro, which doesn't
check for a grave under a raised drawbridge, but currently there are
no levels where such a combination can occur.  The new case probably
does not need to check for a raised drawbridge with a lowered one
under it, because having drawbridges that can lower across one another
would probably be buggy in more important ways than this engrave
message.