Apply Dev Team fix for C343-240, nethack.rankin, Fri Dec 22 04:05:13 2006 +0000
Arming a landmine on the Plane of Air, then setting it off or letting
a monster do so, resulted in a pit.  The fix prevents landmine
explosions from creating a pit on the Planes of Air or Water (in much
the same way that pits are not created when a landmine goes off on a
bridge).  Additionally, the fix also prevents traps (land mines and
bear traps) from being set in such locations in the first place.