Ticket #612, make the fire command confirm when not wielding a needed launcher.
When using the fire command, if the object in the quiver is ammo that
ought to be fired from a launcher, and if the player is not wielding
an appropriate launcher for it, then prompt for confirmation.  Putting
the ammo in your quiver and using the fire command announces your
intention to effect a ranged attack, so if that's not what you're
actually going to be doing, it's very probably a mistake.

The throw command is exempt from the confirmation, because there are
any number of reasons to throw things other than trying to use them as
ammo.  People throw rocks all the time without wielding a sling (e.g.,
it is standard strategy when a hostile monster is behind a boulder in
the Sokoban pit corridor), and I often throw arrows or bolts e.g., at
jellies when not wielding a bow, if they're easier to find a stack of
quickly than daggers or darts while playing a melee character.  The
confirmation would be annoying in these cases, so for throwing I have
chosen to err on the side of letting the user make preventable
mistakes sometimes rather than introducing new annoyances.