Apply Dev Team fix, C343-244, nethack.rankin, Thu Jan 25 06:24:39 2007 +0000
It had been reported on the newsgroup that as a metallivore you could
eat an artifact off the floor that you could not pick up (because it
would evade your grasp).  The fix makes them evade your grasp if you
try it.  Also, eating or attempting to eat a cross-aligned artifact
will blast you now, whether it evades your grasp or not.

Cross-aligned artifacts that blast you still "taste delicious" if you
survive.  The commit log entry raises the question of whether that is
really correct, but the fix does not attempt to address it.