Ticket #480, first batch of new oracles, rumors, and graffiti.
This is low-hanging fruit, consisting of things that were easy to add,
although it would probably be good for someone to run a second set of
eyes over them.

New oracles in this commit advertise magic cancellation (of cloaks and
mithril, in separate oracles); reverse genocide and the making of
dragon scale mail; and alchemy.

New true rumors advertise that not all cloaks are equal (two rumors);
that mithril provides a benefit similar to cloaks; wraith luring; the
effect of age on corpse safety; that zombie corpses are old; tengu
eating; the dangers of fountain eating (two rumors); the connection
between Sokoban and luck; a connection between Mine's End and luck;
the relationship between prayer and wishing; and price identification
for rings and scrolls.

I also threw in some dubious rumors and random graffiti, including
some longer strings of Elbereth (a very minor attempt to make bones
levels very slightly less readily identifiable in some common cases).

The ticket also contains ideas for using major oracle consultations to
reveal game-specific information, such as the locations of certain
branch stairs or which Sokoban prize is available.  This commit does
not attempt to implement any of that.  I'll leave the ticket open.