Apply Dev Team fix for C343-254, nethack.rankin, Fri Feb 9 00:25:47 2007 +0000
If a shapechanger became a mimic, it always appeared as m, never as
an object or furniture.  Similarly, shapechangers that became hiding
monsters did not hide.  This commit allows them to do so.  It also
forces a mimicing shapechanger back out of hiding if it changes
form to something else.

It is possible that this change may introduce bugs.