Allow tombstone to be created if player dies in corridor.
The code in end.c was relying on make_grave to determine whether
the terrain was suitable to be made into a grave or not, and the
code in make_grave was erring on the side of caution.

I audited all places where make_grave is called, and the only one
where this change might have had an unintended effect was in vault.c,
so I added an if clause there to guard against any unintended change.
(I didn't trace the call chain all the way back, but e.g., if paygd is
called before the corridor is converted back to rock, we wouldn't want
to end up with a tombstone engraving in solid rock.)  In all the other
cases, the calling code has already guaranteed that the location under
consideration is in a room, except for the case of wizmode-wishing
for terrain, and I don't see a problem if that works in corridors.

So now if the player dies in a corridor, a tombstone can be created.